Move services to utils
