+++ client/src/utils/textFormat.ts
